Move 'ToStringSerializerTests' to ExtensionsTests class.
